pyproject.toml
[tool.poetry]
name = "whisper-asr-webservice"
version = "1.1.0"
description = "Whisper ASR Webservice is a general-purpose speech recognition webservice."
homepage  = "https://github.com/ahmetoner/whisper-asr-webservice/"
license = "https://github.com/ahmetoner/whisper-asr-webservice/blob/main/LICENCE"
authors = [
    "Ahmet Öner",
    "Besim Alibegovic",
]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.8"
unidecode = "^1.3.4"
uvicorn = { extras = ["standard"], version = "^0.18.2" }
gunicorn = "^20.1.0"
python-multipart = "^0.0.5"
ffmpeg-python = "^0.2.0"
fastapi = "^0.95.1"
openai-whisper = "20230314"
faster-whisper = "^0.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
